pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "leftover-detection"
version = "0.1.0"
description = "AI food waste detection MVP pipeline"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "LeftoverDetection" }]
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
